# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user-transactions": {
      "get": {
        "operationId": "UserTransactionsController_findAll",
        "summary": "Get all user transactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Transactions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/user-transactions/{id}": {
      "get": {
        "operationId": "UserTransactionsController_findOne",
        "summary": "Get user transaction by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Transactions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/user-transactions/user/me": {
      "get": {
        "operationId": "UserTransactionsController_findAllUserTransactions",
        "summary": "Get all user transactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Transactions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/user-address/user/{id}": {
      "get": {
        "operationId": "UserAddressController_findOne",
        "summary": "Get all addresses of a user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Address"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/commission-history": {
      "post": {
        "operationId": "CommissionHistoryController_create",
        "summary": "ADMIN: Create a new commission history",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommissionHistoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create a new commission history operation completed successfully."
          }
        },
        "tags": [
          "Commission History"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "CommissionHistoryController_findAll",
        "summary": "ADMIN: Get all commission histories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all commission histories operation completed successfully."
          }
        },
        "tags": [
          "Commission History"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/commission-history/{id}": {
      "get": {
        "operationId": "CommissionHistoryController_findOne",
        "summary": "Get a commission history by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a commission history by id operation completed successfully."
          }
        },
        "tags": [
          "Commission History"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/commission-history/monthly/{month}/{year}": {
      "get": {
        "operationId": "CommissionHistoryController_findAllByMonthAndYear",
        "summary": "ADMIN: Get all commission histories for a specific month and year",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all commission histories for a specific month and year operation completed successfully."
          }
        },
        "tags": [
          "Commission History"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateCommissionHistoryDto": {
        "type": "object",
        "properties": {
          "userStatusId": {
            "type": "number",
            "description": "ID of the user status",
            "example": 1
          },
          "monthlyCommission": {
            "type": "string",
            "description": "Monthly commission amount",
            "example": "1000000"
          },
          "bonus": {
            "type": "string",
            "description": "Bonus amount",
            "example": "500000"
          },
          "total_sales": {
            "type": "string",
            "description": "Total sales amount",
            "example": "2000000"
          },
          "month": {
            "type": "number",
            "description": "Month of the commission",
            "example": 1
          },
          "year": {
            "type": "number",
            "description": "Year of the commission",
            "example": 2021
          }
        },
        "required": [
          "userStatusId",
          "monthlyCommission",
          "bonus",
          "total_sales",
          "month",
          "year"
        ]
      }
    }
  }
}